Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/navigation-timing/nav2_test_redirect_xserver-expected.txt

Blocked access to external URL http://www.localhost:8800/common/redirect.py?location=http://localhost:8800/navigation-timing/resources/blank_page_green.html
CONSOLE MESSAGE: line 16: TypeError: undefined is not an object (evaluating 'performanceNamespace.getEntriesByType("navigation")[0].type')
Description

This test validates the values of the window.performance.getEntriesByType("navigation")[0].redirectCount and the window.performance.getEntriesByType("navigation")[0].redirectStart/End times for a cross-origin server side redirect navigation.



FAIL Navigation Timing 2 WPT TypeError: undefined is not an object (evaluating 'performanceNamespace.getEntriesByType("navigation")[0].type')
